Passing in/out channels between modules when creating the model.
Passing in/out channels between modules when creating the model. This simplifies the code to compute the input channels for feature extractors and makes the predictors independent of the backbone architectures.
* Passed in_channels to rpn and head builders.
* Set out_channels to model modules including backbone and feature extractors.
* Moved cfg.MODEL.BACKBONE.OUT_CHANNELS to cfg.MODEL.RESNETS.BACKBONE_OUT_CHANNELS as it is not used by all architectures. Updated config files accordingly.

For new architecture modules, the return module needs to contain a field called 'out_channels' to indicate the output channel size.